Refactor to use MlflowRegistry class
